backup_20250323_205738/requirements.txt
numpy>=1.20.2
opencv-python>=4.2.0.32
opencv-contrib-python>=4.2.0.32
streamlit>=1.0.0
Pillow>=8.2.0
tensorflow>=2.7.2
scikit-learn>=0.22.1
matplotlib>=3.3.3
watchdog>=4.0.2